// PennyTracker.Frontend/src/enums/TransactionCategory.ts
export enum TransactionCategory {
  // Income Categories
  Salary = 'Salary',
  Bonus = 'Bonus',
  Investments = 'Investments',
  Freelance = 'Freelance',
  Gift = 'Gift',
  Refund = 'Refund',

  // Expense Categories
  Food = 'Food',
  Rent = 'Rent',
  Entertainment = 'Entertainment',
  Utilities = 'Utilities',
  Transportation = 'Transportation',
  Healthcare = 'Healthcare',
  Education = 'Education',

  // Others / Common
  Miscellaneous = 'Miscellaneous'
}